# Conflicts:
#	src/main/java/org/example/CMVCalculator.java
#	src/main/java/org/example/Decide.java
#	src/test/java/org/example/CMVCalculatorTest.java

+++ src/main/java/org/example/Decide.java
@@ -1,7 +1,7 @@
 package org.example;
 
 public class Decide {
-    public static String decide(float[][] points, Parameters parameters, OperationType[][] lcm, boolean[] puv) {
+    public static String decide(double[][] points, Parameters parameters, OperationType[][] lcm, boolean[] puv) {
         var cmv = CMVCalculator.calculateCMV(points, parameters);
         var pum = PUMCalculator.calculatePUM(cmv, lcm);
         var fuv = FUVCalculator.calculateFUV(pum, puv);
